Update printingtoolsng-help.html (fr)
Hello Christopher,
I hope you're well.
This is an update of the French help file for version 3.1.0 because the version of the file you used is obsolete. On April 20, 2023, I sent you a pull request to update the file, but you didn't use it. Could this have been a handling error?
Anyway, I had a problem translating the part about adding the button to the message header toolbar. In fact, in my Thunderbird 115 on Windows 7, it's not possible to add a button via More > Customize. To do this, I have to use another addon: "Message Header Toolbar Customize".